create homeViewModel
