--- Pethouse/HomeView.swift
@@ -1,24 +1,0 @@
-//
-//  ContentView.swift
-//  Pethouse
-//
-//  Created by Gustavo Souto Pereira on 02/07/25.
-//
-
-import SwiftUI
-
-struct HomeView: View {
-    var body: some View {
-        VStack {
-            Image(systemName: "globe")
-                .imageScale(.large)
-                .foregroundStyle(.tint)
-            Text("Hello, world!")
-        }
-        .padding()
-    }
-}
-
-#Preview {
-    HomeView()
-}
