+++ Pethouse/View/Home/HomeViewModel.swift
@@ -1,0 +1,12 @@
+//
+//  HomeViewModel.swift
+//  Pethouse
+//
+//  Created by Thiago de Jesus on 02/07/25.
+//
+
+import Foundation
+@Observable
+class HomeViewModel {
+    
+}
